[fix] threading, intermittent failure
-preferring to push and forget work from the service's 'add_links'
request handler, use of 'asyncio.ensure_future' to add the coroutine
awaitable to the main async loop resulted in:

'ERROR:asyncio:Task exception was never retrieved'

which following debug mode led to uncover:

RuntimeError: Non-thread-safe operation invoked on an event loop other
than the current coroutine

which persisted even when pushing the correct main async loop (as
opposed to the service's bespoke (tornado) 'ioloop'

however, 'asyncio.ensure_future' wraps the awaitable coroutine in a
Task, and of note is the description of task:

'A Future-like object that runs a Python coroutine. Not thread-safe.'

the spider web service is running in its own thread. switching to
'asyncio.run_coroutine_threadsafe' to schedule the coroutine in the
main thread's async loop was the remedy